fs_books/requirements.txt
###  python example of connect to GCP firestore and storage
## run as ./bin/gunicorn -b :8080 main:app

Flask==3.1.1
gunicorn==23.0.0
google-cloud-logging==3.12.1
google-cloud-firestore==2.21.0
google-cloud-storage==2.17.0